updated snippet names with correct order
